# src/ServiceFabricPubSub/ClientApp/ClientApi/swagger_admin_api.json
{"swagger":"2.0","info":{"version":"v2","title":"PubSubAdminApi"},"host":"localhost:8979","schemes":["http"],"paths":{"/api/Subscriber":{"get":{"tags":["Subscriber"],"operationId":"Subscriber_GetSubscribers","consumes":[],"produces":["application/json","text/json"],"parameters":[{"name":"tenantId","in":"query","required":true,"type":"string"},{"name":"topicName","in":"query","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}},"put":{"tags":["Subscriber"],"operationId":"Subscriber_AddSubscriber","consumes":[],"produces":["application/json","text/json"],"parameters":[{"name":"tenantId","in":"query","required":true,"type":"string"},{"name":"topicName","in":"query","required":true,"type":"string"},{"name":"subscriberName","in":"query","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}},"delete":{"tags":["Subscriber"],"operationId":"Subscriber_DeleteSubscriber","consumes":[],"produces":["application/json","text/json"],"parameters":[{"name":"tenantId","in":"query","required":true,"type":"string"},{"name":"topicName","in":"query","required":true,"type":"string"},{"name":"subscriberName","in":"query","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/api/Tenants":{"get":{"tags":["Tenants"],"operationId":"Tenants_ResetKeys","consumes":[],"produces":["application/json","text/json"],"parameters":[{"name":"tenantId","in":"query","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}}}},"put":{"tags":["Tenants"],"operationId":"Tenants_CreateTenant","consumes":[],"produces":["application/json","text/json"],"parameters":[{"name":"TenantName","in":"query","required":true,"type":"string"},{"name":"AppVersion","in":"query","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}}}}},"/api/Topics":{"get":{"tags":["Topics"],"operationId":"Topics_GetTopics","consumes":[],"produces":["application/json","text/json"],"parameters":[{"name":"tenantId","in":"query","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}},"put":{"tags":["Topics"],"operationId":"Topics_CreateTopic","consumes":[],"produces":["application/json","text/json"],"parameters":[{"name":"tenantId","in":"query","required":true,"type":"string"},{"name":"TopicName","in":"query","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}}}},"delete":{"tags":["Topics"],"operationId":"Topics_DeleteTopic","consumes":[],"produces":["application/json","text/json"],"parameters":[{"name":"tenantId","in":"query","required":true,"type":"string"},{"name":"TopicName","in":"query","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}}}}}},"definitions":{}}